[build-script] Transform IndexStoreDB and SorcekitLSP to use ProductBuilder.
Create a new helper builder for IndexStoreDB and SourceKitLSP called
BuildScriptHelperBuilder (because the invoked script is called
build-script-helper.py). Move most of the code that existed in a
function in indexstore.db into that helper class.

Create small subtypes of the helper paramtrized for IndexStoreDB and
SourceKitLSP.

Modify the main build-script to invoke the new build instead of the
Product class directly. This way, when all the products use builders
to be build, all of them will use the same Python invocations, but
some of them will use build-script-impl, and others will use
build-script-helper.py.