[TF] Further cleanup of CompilerRuntime and friends (#24938)
* Remove _ExecutionMode (TF-517)

  * TF_CreateConfig.enable_xla_compilation is now hardcoded to 0.
  * TPU tests now unconditionally fail instead of pretending to work.

* Remove enableTPU, enableGPU and enableCPU

These functions have empty bodies, but they have pretty high-profile
names which gives an impression that they do something important.

With that in mind, I think it would be best to remove these functions
without any deprecation notice because there aren't any comparable
replacements at the moment.

* Remove obsolete remnants of the TPU testing infrastructure (TF-520)

This commit removes the remnants of the obsolete TPU testing
infrastructure. In the nearest future, we are planning to restore it
within the new tracing runtime (https://bugs.swift.org/browse/TF-520).

* Remove obsolete remnants of the TPU runtime infrastructure (TF-519)

This commit removes the remnants of the obsolete TPU runtime
infrastructure. In the nearest future, we are planning to restore it
within the new tracing runtime (https://bugs.swift.org/browse/TF-519).