[5.1] IRGen: Fix getLocalSelfMetadata
Don't use the beginning of the basic block as insertion point if 'local
self' is an llvm::Instruction (i.e a cast from the argument llvm::Value
has been insterted) rather use the 'local self' value as insertion point.

Fixes a dominance issue because we inserted at the beginning of the
basic block.

rdar://51186963