[AutoDiff] TF-508: Temporary Hack Fix (#25115)
[JIRA Ticket TF-508](https://bugs.swift.org/browse/TF-508)

This ticket should not be closed, as this PR is a temporary hack to fix the problem and unblock me on `SIMD` vectors.

The problem has to do with not being able to resolve `Self` for the actual underlying struct type. For example, in the reproducer in the JIRA ticket, it doesn't know about `MyStruct`. So here, we take a look at the first type, which we are assuming is `Self`, but more specfically the struct which lets us successfully check the requirements.