Sema: Remove dead code concerning 'default constructible' classes
We used to try to emit a 'default initializer' in a class with a superclass,
as long as the superclass was 'default initializable', meaning it had at
least one designated initializer where all parameters had a default
expression.

However this code path was never taken, because the designated initializer
inheritance mechanism supercedes it. Probably it became dead once we
implemented inheritance of initializers with default arguments.