Sema: Don't emit stubs to override inaccessible superclass initializers
There was a behavioral difference between binary modules and textual interfaces.
Since a binary module includes information about all members, including private
and internal members, we would emit overrides for all designated initializers
when subclassing a class, including those we cannot access.

This was problematic because then we reference the superclass initializer's
method descriptor, which had to be forced to have public linkage so that the
subclass could reference it.

However a textual interface only includes public members so any such
initializers were simply dropped. The IRGen hack is thus no longer necessary
when textual interfaces are used.

In reality neither behavior is correct because the presence of inaccessible
initializers should inhibit the inheritance of any designated initializer;
however I'm going to try to fix that separately since it is a source breaking
change and thus needs to be staged in as a warning. The latter fix is
tracked by <rdar://problem/51249311>.