[windows] Fix module-cache-bad-version.swift in Windows.
In Windows, the expansion of globs is not perform by the shell, and
might happen or not in each process. Seems that some versions of `test`
(like the one used in CI) do their own expansion. Since the test is
testing for the inexistence of a file, the asterisk ends up in the
process argv, then the expansion happen, which finds no files, and test
fails saying that -f needs an argument. This doesn't happen in Unix
because the asterisk is not expanded in process.

Quoting or other tricks were defeated by lit own parsing of the
commands.

The solution involves using Python own globbing to find out if there's a
file that matches the glob or not.