SILOptimizer: Simplify dead method elimination
We used to look at the vtable entry's linkage to determine if the
vtable entry was dead, in case the vtable entry linkage was more
visible than the method being overridden.

This hack is no longer needed now that a method override more
visible than its base introduces a new vtable entry.