tests: improve resilience test helper for Windows
Windows does not support the concept of RPATH.  As a result, we must
change the working directory for the test.  This extends the support
for the tests to work with PE/COFF on Windows.  Use the newly minted
`self.triple` property to drive the host detection.  This means that
the test coverage on Windows is able to test most of the resilience
functionality.