test: add a workaround for `rth` invocation
The `rth` tool is a python script and relied on the execution via the
shebang.  However, not all targets support the shebang to invoke the
interpreter.  Explicitly launch the resilience test helper with the
interpreter (python).  Ideally, we would use the `%rth` substitution,
but that does not seem to function for the nested case here.