SILGen: make `#dsohandle` work properly on Windows
Windows has a special symbol `__ImageBase` whcih provides the constant
value of the base of the image.  This is roughly equivalent to the
`__dso_handle` on the ELF and MachO targets.  Use this to construct the
value for the `#dsohandle` rather than attempting to use the
non-existent symbol `__dso_handle`.  This fixes the
dsohandle-multi-module validation test on Windows.