[AutoDiff] Partially fix AD control flow memory leaks (#25294)
This PR partially fixes the incorrect omission of memory cleanup in `AdjointEmitter`.

Cleanups should be created per basic block in an adjoint function, starting with its arguments. Like function parameter arguments in the adjoint entry block, every phi argument except the pullback struct in a non-entry adjoint block also needs to possess a cleanup. We let the existing adjoint emission logic accumulate the cleanups until we are about to branch, where we disable the top-level cleanup for arguments in the terminator instruction and apply all of their child cleanups recursively. This logic is more general than the original single-block differentiation logic, and we should refactor `AdjointEmitter::run()` to eliminate special logic for adjoint entry and exit blocks as much as possible.

This PR also unifies the argument order of the adjoint entry block with that of other adjoint blocks, making the pullback struct always the last argument. This helps us reduce special logic for the adjoint entry block later.