[Frontend] Add an option to emit XCTest methods to a file
This adds an option "-emit-xctest-methods-path" to the driver and
frontend for emitting a list of XCTest methods in a JSON file. This will
be used by SwiftPM (and potentially other build systems) to
automatically generate LinuxMain.swift file during the build process so
users don't have to manually maintain the file on Linux.

This is a much easier solution than other alternatives like invoking
SourceKit during a build. Eventually, XCTest on Linux should be able to
use runtime reflection to query the test methods but we don't have that
yet.