stdlib: Add backward deployment versions for the dynamic-replacement runtime functions.
This will allow backward deployment to a swift 5.0 runtime.
It also needs a different mechanism to pass the "implicit" boolean parameter in the thread local storage: Instead storing it in a boolean field, we use bit 0 of the AccessSet Head.
This avoids changing the layout of the thread local storage and makes it compatible with the swift 5.0 runtime.

rdar://problem/51601233