Disallow `dynamic` on native Swift declarations under library-evolution mode
The current ABI of dynamic is not what we want for certain declarations.

e.g

dynamic var x : Int {
  willset {
  }
}

Is exposed in the .swiftinterface file as

dynamic var x : Int {
  get {}
  set {}
}

However, the actual implementation is marking the willSet only dynamic.

var x : Int {
  dynamic willSet() {}
  get {}
  set {}
}

To be able to rely on `dynamic` in the future we disable dynamic (for
non-@objc) under library-evolution for now until we have made sure that
the ABI is proper.

rdar://51678074