[cast-optimizer] Fix a leak in the cast optimizer when we perform load promotion and fail.
Specifically, in this part of the cast optimizer we are trying to optimize casts
that can be done in two parts. As an example consider: NSObject ->
Array<Any>. In this case, we first cast from NSObject -> NSArray and then try to
conditionally bridge to Array<Any> from NSArray.

The problem is we did not destroy the NSObject correctly if the first cast
failed. I couldn't figure out how to create an actual swift test case that
produces this problem since we are pretty conservative about triggering this
code path. But in SIL it is pretty easy and in ossa, we trigger the ownership
verifier.

This is another victory for the ownership verifier!

rdar://51753580
(cherry picked from commit c5e9fc286a9f6d5774ad7d8f7831a0f84a80fd71)

Conflicts:
	lib/SILOptimizer/Utils/CastOptimizer.cpp
	test/SILOptimizer/constant_propagation_objc.sil
	test/SILOptimizer/constant_propagation_ownership_objc.sil

The conflicts came from my having on master extracted out the piece of code
where this is done into a helper function.