[WIP] Constrain KeyPathExprs to never be PartialKeyPaths
[WIP: Regresses diagnostics for testSubtypeKeypathProtocol subtest]

KeyPathExprs should never directly generate an AnyKeyPath or PartialKeyPath; they should always generate either a KeyPath or one of its subclasses and then upcast it. simplifyKeyPathConstraint() currently does this by erroring out when the constraint solver tries to make it a PartialKeyPath, counting on the constraint solver as a whole to try its subclasses. That could fail if the constraint solver never tries any of them.

This commit changes simplifyKeyPathConstraint() so that, when it re-adds itself because there are overloads that haven’t yet been resolved, it also constrains its type to be KeyPath or a subtype if the typechecker seems to need additional guidance.